+++ cmake/XSTestsLibraries.cmake
@@ -5,7 +5,7 @@ foreach(source_fixture MissingMain NonLiteralMain OutOfRangeMain OutOfRangeByteM
                        ParameterizedMain WrongReturnMain UnknownCallMain
                        WrongCallArityMain NonLongReturnCallMain
                        BoolCallAsLongMain UnitCallAsLongMain InvalidLogicalOperands ImmutableLocalReassignment
-                       MatchMissingElse MatchPatternTypeMismatch)
+                       MatchMissingElse MatchPatternTypeMismatch RecursiveDataParameter)
   add_test(NAME source_native_invalid_${source_fixture} COMMAND xs build -file
                                                            ${XS_SOURCE_NATIVE_FIXTURE_DIR}/${source_fixture}.xs)
   set_tests_properties(source_native_invalid_${source_fixture} PROPERTIES TIMEOUT 5 WILL_FAIL TRUE)